+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Upgrade the `org.realityforge.proton` artifacts to version `0.40`.
+* Stop generating errors if types are annotated with `@Named`, `@Typed` or `@Eager` and they have a provider annotation.
 
 ### [v0.01](https://github.com/sting-ioc/sting/tree/v0.01) (2020-02-13) · [Full Changelog](https://github.com/sting-ioc/sting/compare/9e796d0e5c44bee98107f3e65bd394d41bbe07c7...v0.01)